[core] Remove `gcs_rpc_server.h` dependency from `GcsWorkerManager` (#55954)
Splitting gRPC service interface from implementation.

---------
