+++ src/ray/rpc/gcs/gcs_rpc_server.h
@@ -142,11 +142,6 @@ namespace rpc {
                       HANDLER,                   \
                       RayConfig::instance().gcs_max_active_rpcs_per_handler())
 
-#define WORKER_INFO_SERVICE_RPC_HANDLER(HANDLER) \
-  RPC_SERVICE_HANDLER(WorkerInfoGcsService,      \
-                      HANDLER,                   \
-                      RayConfig::instance().gcs_max_active_rpcs_per_handler())
-
 #define PLACEMENT_GROUP_INFO_SERVICE_RPC_HANDLER(HANDLER) \
   RPC_SERVICE_HANDLER(PlacementGroupInfoGcsService,       \
                       HANDLER,                            \
@@ -361,68 +356,6 @@ class NodeResourceInfoGrpcService : public GrpcService {
   NodeResourceInfoGcsService::AsyncService service_;
   /// The service handler that actually handle the requests.
   NodeResourceInfoGcsServiceHandler &service_handler_;
-};
-
-class WorkerInfoGcsServiceHandler {
- public:
-  virtual ~WorkerInfoGcsServiceHandler() = default;
-
-  virtual void HandleReportWorkerFailure(ReportWorkerFailureRequest request,
-                                         ReportWorkerFailureReply *reply,
-                                         SendReplyCallback send_reply_callback) = 0;
-
-  virtual void HandleGetWorkerInfo(GetWorkerInfoRequest request,
-                                   GetWorkerInfoReply *reply,
-                                   SendReplyCallback send_reply_callback) = 0;
-
-  virtual void HandleGetAllWorkerInfo(GetAllWorkerInfoRequest request,
-                                      GetAllWorkerInfoReply *reply,
-                                      SendReplyCallback send_reply_callback) = 0;
-
-  virtual void HandleAddWorkerInfo(AddWorkerInfoRequest request,
-                                   AddWorkerInfoReply *reply,
-                                   SendReplyCallback send_reply_callback) = 0;
-
-  virtual void HandleUpdateWorkerDebuggerPort(UpdateWorkerDebuggerPortRequest request,
-                                              UpdateWorkerDebuggerPortReply *reply,
-                                              SendReplyCallback send_reply_callback) = 0;
-
-  virtual void HandleUpdateWorkerNumPausedThreads(
-      UpdateWorkerNumPausedThreadsRequest request,
-      UpdateWorkerNumPausedThreadsReply *reply,
-      SendReplyCallback send_reply_callback) = 0;
-};
-
-/// The `GrpcService` for `WorkerInfoGcsService`.
-class WorkerInfoGrpcService : public GrpcService {
- public:
-  /// Constructor.
-  ///
-  /// \param[in] handler The service handler that actually handle the requests.
-  explicit WorkerInfoGrpcService(instrumented_io_context &io_service,
-                                 WorkerInfoGcsServiceHandler &handler)
-      : GrpcService(io_service), service_handler_(handler){};
-
- protected:
-  grpc::Service &GetGrpcService() override { return service_; }
-
-  void InitServerCallFactories(
-      const std::unique_ptr<grpc::ServerCompletionQueue> &cq,
-      std::vector<std::unique_ptr<ServerCallFactory>> *server_call_factories,
-      const ClusterID &cluster_id) override {
-    WORKER_INFO_SERVICE_RPC_HANDLER(ReportWorkerFailure);
-    WORKER_INFO_SERVICE_RPC_HANDLER(GetWorkerInfo);
-    WORKER_INFO_SERVICE_RPC_HANDLER(GetAllWorkerInfo);
-    WORKER_INFO_SERVICE_RPC_HANDLER(AddWorkerInfo);
-    WORKER_INFO_SERVICE_RPC_HANDLER(UpdateWorkerDebuggerPort);
-    WORKER_INFO_SERVICE_RPC_HANDLER(UpdateWorkerNumPausedThreads);
-  }
-
- private:
-  /// The grpc async service object.
-  WorkerInfoGcsService::AsyncService service_;
-  /// The service handler that actually handle the requests.
-  WorkerInfoGcsServiceHandler &service_handler_;
 };
 
 class PlacementGroupInfoGcsServiceHandler {
@@ -665,7 +598,6 @@ class InternalPubSubGrpcService : public GrpcService {
 using JobInfoHandler = JobInfoGcsServiceHandler;
 using ActorInfoHandler = ActorInfoGcsServiceHandler;
 using NodeResourceInfoHandler = NodeResourceInfoGcsServiceHandler;
-using WorkerInfoHandler = WorkerInfoGcsServiceHandler;
 using PlacementGroupInfoHandler = PlacementGroupInfoGcsServiceHandler;
 using InternalKVHandler = InternalKVGcsServiceHandler;
 using InternalPubSubHandler = InternalPubSubGcsServiceHandler;
